Move sessions hooks/state/api to its own lib
