+++ client/src/lib/sessions/api/session.ts
@@ -1,5 +1,5 @@
 import {ExerciseState, Session} from '../../../../../shared/src/types/Session';
-import apiClient from '../../../lib/apiClient/apiClient';
+import apiClient from '../../apiClient/apiClient';
 
 const SESSIONS_ENDPOINT = '/sessions';
