+++ client/src/lib/sessions/hooks/usePinnedSessions.ts
@@ -1,9 +1,9 @@
 import dayjs from 'dayjs';
 import {useCallback, useMemo} from 'react';
 import {Session} from '../../../../../shared/src/types/Session';
-import useLogSessionMetricEvents from '../../../routes/Session/hooks/useLogSessionMetricEvents';
-import useUserState from '../state/state';
-import useCurrentUserState from './useCurrentUserState';
+import useLogSessionMetricEvents from './useLogSessionMetricEvents';
+import useUserState from '../../user/state/state';
+import useCurrentUserState from '../../user/hooks/useCurrentUserState';
 
 const usePinnedSessons = () => {
   const userState = useCurrentUserState();
